refactor(types): menu
